kali-rsyslog.conf
Installed rsyslog on a Kali Linux endpoint to safely transmit system and firewall logs over the network to an endpoint Graylog SIEM of centralized logs. Applied the UDP/TCP log forwarding on port 1514 and tested the connection and ingested the logs successfully on the Graylog to monitor and analyze in real-time.